feat: Support XWiki Format
  * Integrate new formats directly in ttkit file
  * Move tests to test_formats
