--- weblate/formats/tests/xwiki/data/xwiki.properties
@@ -1,74 +1,0 @@
-# ---------------------------------------------------------------------------
-# See the NOTICE file distributed with this work for additional
-# information regarding copyright ownership.
-#
-# This is free software; you can redistribute it and/or modify it
-# under the terms of the GNU Lesser General Public License as
-# published by the Free Software Foundation; either version 2.1 of
-# the License, or (at your option) any later version.
-#
-# This software is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-# Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public
-# License along with this software; if not, write to the Free
-# Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-# 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-# ---------------------------------------------------------------------------
-
-###############################################################################
-# XWiki Core localization
-#
-# This contains the translations of the module in the default language
-# (generally English).
-# 
-# Translation key syntax:
-#   <short top level project name>.<short module name>.<propertyName>
-#   where:
-#   * <short top level project name> = top level project name without the "xwiki-" prefix,
-#                                     for example: commons, rendering, platform, enterprise, manager, etc
-#   * <short module name> = the name of the Maven module without the <short top level project name> prefix,
-#                           for example: oldcore, scheduler, activitystream, etc
-#   * <propertyName> = the name of the property using camel case,
-#                      for example updateJobClassCommitComment
-#
-# Comments: it's possible to add some detail about a key to make easier to
-#   translate it by adding a comment before it. To make sure a comment is not
-#   assigned to the following key use at least three sharps (###) for the comment
-#   or after it.
-# 
-# Deprecated keys:
-#   * when deleting a key it should be moved to deprecated section at the end
-#     of the file (between #@deprecatedstart and #@deprecatedend) and associated to the
-#     first version in which it started to be deprecated
-#   * when renaming a key, it should be moved to the same deprecated section
-#     and a comment should be added with the following syntax:
-#     #@deprecated new.key.name
-#     old.key.name=Some translation
-###############################################################################
-
-job.question.button.confirm=Confirm the operation {0}
-job.question.button.cancel=Cancel
-job.question.notification.canceling=Canceling with ''
-job.question.notification.answering=Answering to test '' {3}
-job.question.\ escapedSpace.answering=Property with escaped space
-###############################################################################
-## Deprecated
-## Note: each element should be removed when the last branch using it is no longer supported
-###############################################################################
-
-## Used to indicate where deprecated keys start
-
-#@deprecatedstart
-
-#######################################
-## until 10.1
-#######################################
-
-job.status.success=Done.
-job.status.canceled=Canceled.
-job.status.error=Failed.
-#@deprecatedend
-
