+++ weblate/trans/tests/data/XWikiFullPage.xml
