+++ weblate/trans/tests/data/XWikiPageProperties.xml
