+++ weblate/trans/tests/data/xwiki.properties
@@ -49,15 +49,26 @@
 #     old.key.name=Some translation
 ###############################################################################
 
-IGNORE=Ignore
-IGNOREALL=Ignore All
-ADD=Add to Dictionary
-REPLACE=Change
-REPLACEALL=Change All
-CANCEL=Cancel
-SUGGESTIONS=Suggestions
-INVALIDWORD=Not in Dictionary
-COMPLETED=Spellcheck completed.
-ADDWORD_1=Add
-ADDWORD_2=to dictionary?
-ADDWORD_3=Add Word?
+job.question.button.confirm=Confirm the operation {0}
+job.question.button.cancel=Cancel
+job.question.notification.canceling=Canceling with ''
+job.question.notification.answering=Answering to test '' {3}
+job.question.\ escapedSpace.answering=Property with escaped space
+###############################################################################
+## Deprecated
+## Note: each element should be removed when the last branch using it is no longer supported
+###############################################################################
+
+## Used to indicate where deprecated keys start
+
+#@deprecatedstart
+
+#######################################
+## until 10.1
+#######################################
+
+job.status.success=Done.
+job.status.canceled=Canceled.
+job.status.error=Failed.
+#@deprecatedend
+
